requirements.txt
Flask==3.0.0
pandas==2.0.3
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.8
Werkzeug==3.0.1
python-dotenv==1.0.0
requests

